Tooltip components and util moved to subfolder
