+++ components/tooltip/index.tsx
@@ -1,0 +1,2 @@
+export { TOOLTIP_MODE, POINTS_TO } from './TooltipPositioned';
+export { default as TooltipWrapper } from './Tooltip';
